requirements.txt
streamlit>=1.28.0
supabase>=1.0.0
pandas>=1.5.0
plotly>=5.0.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
PyJWT>=2.6.0
requests>=2.28.0
python-multipart>=0.0.5
email-validator>=1.3.0
